+++ src/Products/PloneMeeting/indexes.py
@@ -139,8 +139,8 @@ def reviewProcessInfo(obj):
       and the item review_state so it can be queryable in the catalog.
     """
     item_state = obj.query_state()
-    return '%s__reviewprocess__%s' % (
-        obj.adapted()._getGroupManagingItem(item_state, theObject=False), item_state)
+    return ['%s__reviewprocess__%s' % (org_uid, item_state)
+            for org_uid in obj._getGroupsManagingItem(item_state)]
 
 
 @indexer(IMeetingItem)